requirements.txt
openai>=1.51.0
python-dotenv>=1.0.1
numpy>=1.23.0
faiss-cpu>=1.7.4

# Local embeddings
sentence-transformers>=2.2.0
torch>=2.0.0
